backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
yfinance==0.2.48
requests==2.32.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
curl_cffi==0.6.0
feedparser==6.0.11
gunicorn==21.2.0
